move `balance` to `const` del `texts`
