--- bot/consts/texts/__init__.py
@@ -1,1 +1,0 @@
-from .balance import *
